chore: fixing up tests and extracting adminThumnail to its own file
